Add concurrent username check
UniqueUsername agent  started with  Accounts context
supervisor  to claim  usernames  to avoid  situation
when the same username  would be registered twice at
(almost)  the  same time.  The  DB  is checked  with
"check_username", but  even with  strong concurrency
turned on  if, for  example, multiple admins  try to
add the  same user, one  will succeed and  the other
one will crash  the system and events  would have to
be  cleaned  up.  But  because the  event  store  is
immutable,  that is  a  bit no-no,  so an  "error/3"
callback  would have  to be  added to  the Projector
module to ignore "Ecto.ConstraintError"s.

See "error/3" example at https://github.com/commanded/commanded-ecto-projections#error3-callback