Change read model credential system
The notion that a user can have multiple credentials
is  solid I  think (see  also 2018-10-14_2339),  but
there  is no  reason  for a  user  to have  multiple
`username_password` credentials  (hence the `Schema`
changes from `has_many` to `has_one`).

I  am also  still prone  to confuse  READ and  WRITE
sides:  the `users_credentals`  table  was about  to
become very  generic with  maps to hold  the payload
of  credential  types  (such as  the  CQRS  commands
do),  but that  is  insane. The  logical thing  with
relation databases would be to hold a table for each
credential  type  and  project  events  of  specific
credential types into  them (see 2019-01-30_0628 for
more  on this).  If  I  understand correctly,  table
relationships are implicit  in RDBMS' (`references`)
compared  to `Ecto.Schemas`  (`has_many`, `has_one`,
etc.).

```text
____
    |
user|<-- username_password_credentials
____|<-- facebook_credentials
```

Entities having a credential `:type` field:
+ `AddUsernamePasswordCredential` command
  at the moment)
+ `Credential` aggregate

... and that do not:
+ Credential events
+ `Read.Schemas` credentials (`username_password_credentials`
  at the moment)

CQRS commands  update the streams'  (i.e., aggregate
instance processes)  state and they  are polymorphic
(it wouldn't be  very good to have  an aggregate for
each credential type). The `:type` field is probably
only important  on new credential  creation, because
even  though  the  commands'  name  is  descriptive,
the aggregate  template is type-agnostic  (hence the
payload  field).  Therefore   when  creating  a  new
aggregate  instance, it  stores the  credential type
from  the  command,  but   later  commands  will  be
identified  by ID  and they  probably won't  need to
expose this information.

Event names are descriptive as well, and can be used
to project them into the right tables, therefore the
`:type` field is not needed.

The RDBMS tables omit the field on the same basis.